WIP: change team members into components #351
